+++ CHANGELOG.md
@@ -3,9 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.2.0] - 2020-06-18
 ### Added
+* `utils.mapping.human_symbol_to_human_ens()`
 ### Changed
+* anova no longer uses multiprocessing. This should decrease memory usage significantly.
+* es_mu and es_sd now have index column name: "gene".
+* mapping module naming and docs
 ### Deprecated
 ### Removed
 ### Fixed